pyproject.toml
[tool.poetry]
name = "per-point-pate"
version = "0.2.0"
description = "This repository provides an implementation of PATE, as well as extensions of it to enable the personalization of privacy preservation, and a framework for experimentation. The code is designed in an object-oriented manner."
authors = []
readme = "README.md"
packages = [{include = "per_point_pate"}]


[tool.poetry.scripts]
ppp-generate-prms = "per_point_pate.hprms_search:generate_prms"
ppp-hprms-search = "per_point_pate.hprms_search:hprms_search"
ppp-run = "per_point_pate.main:main"
ppp-run-experiment-set = "per_point_pate.main:run_experiment_set"




[tool.poetry.dependencies]
python = ">=3.9, <3.10"
#tensorflow = "^2.9.1"
numpy = "^1.22.4"
matplotlib = "^3.5.2"
pandas = "^1.4.2"
torch = "^1.11.0"
torchvision = "^0.12.0"
click = "^8.1.3"
sklearn = "^0.0"
seaborn = "^0.11.2"
scipy = "^1.8.1"
loguru = "^0.6.0"
tqdm = "^4.64.0"

[tool.poetry.group.dev.dependencies]
PyYAML = "^6.0"
bpython = "^0.22.1"
yapf = "^0.32.0"
black = "^22.3.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
